Change duration string to what it really should say
The 'duration' string is currently saying 'Duration (minutes)'.
However, the table rows on view.php and in the Moodle mobile app where this string is used push the meeting duration through format_time() which renders the duration in a nice format like '1 hour'. Thus, the 'minutes' hint is misleading and removed from the string.